{"schema":"decodex/commit/1","summary":"Split MCP core tool modules","authority":"manual"}
